Create README.md
While Python has libraries and functions and modules that give you the day of the week for any date in just a couple of lines, can you write your own function using just math in order to do this task? I did it as an exercise in Python programming. I just started Python and and self-studying. Comments welcome.